requirements.txt
snowflake-connector-python>=3.6
pandas>=2.0
python-dotenv>=1.0
pyyaml>=6.0
pytest>=8.0
